requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
clickhouse-connect==0.10.0
colorama==0.4.6
Django==5.2.10
django-filter==25.2
djangorestframework==3.16.1
gunicorn==24.1.1
idna==3.11
iniconfig==2.3.0
interrogate==1.7.0
lz4==4.4.5
pluggy==1.6.0
psycopg2-binary==2.9.11
py==1.11.0
pyarmor==9.2.3
pyarmor.cli.core==8.1.0
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-django==4.11.1
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
ruff==0.14.14
setuptools==80.10.1
sqlparse==0.5.5
tabulate==0.9.0
urllib3==2.6.3
wheel==0.46.3
zstandard==0.25.0
